+++ lib/core/engine_clang.cpp
@@ -35,7 +35,8 @@
 #include <boost/optional.hpp>
 #include <boost/range/adaptor/transformed.hpp>
 
-#include <fstream>
+#include <just/file.hpp>
+
 #include <memory>
 #include <algorithm>
 #include <iterator>
@@ -345,6 +346,70 @@ namespace
     return args;
   }
 
+  template <class InputIt>
+  std::set<boost::filesystem::path> split_file_list(InputIt begin_,
+                                                    InputIt end_)
+  {
+    std::set<boost::filesystem::path> result;
+    bool escaped = false;
+    std::string current;
+    for (InputIt i = begin_; i != end_; ++i)
+    {
+      if (escaped)
+      {
+        switch (*i)
+        {
+        case ' ':
+        case '\t':
+          current += *i;
+          break;
+        case '\r':
+        case '\n':
+          if (!current.empty())
+          {
+            result.insert(current);
+            current.clear();
+          }
+          break;
+        default:
+          current += '\\';
+          current += *i;
+        }
+        escaped = false;
+      }
+      else
+      {
+        switch (*i)
+        {
+        case ' ':
+        case '\r':
+        case '\n':
+        case '\t':
+          if (!current.empty())
+          {
+            result.insert(current);
+            current.clear();
+          }
+          break;
+        case '\\':
+          escaped = true;
+          break;
+        default:
+          current += *i;
+        }
+      }
+    }
+    if (escaped)
+    {
+      current += '\\';
+    }
+    if (!current.empty())
+    {
+      result.insert(current);
+    }
+    return result;
+  }
+
   template <bool UseInternalTemplight>
   class engine_clang : public iface::engine
   {
@@ -593,6 +658,26 @@ namespace
       }
 
       return result;
+    }
+
+    virtual std::set<boost::filesystem::path>
+    files_included_by(const std::string& exp_) override
+    {
+      const boost::filesystem::path d_file = _internal_dir / "out.d";
+      const data::process_output output =
+          run_clang(_clang_binary, {"-MD", "-MF", d_file.string(), "-E"}, exp_);
+
+      if (output.exit_code() == data::exit_code_t(0) && exists(d_file))
+      {
+        const auto deps = just::file::read(d_file.string());
+        const auto start = std::find(begin(deps), end(deps), ':');
+        if (start != end(deps))
+        {
+          return split_file_list(start + 1, end(deps));
+        }
+      }
+      throw std::runtime_error("Error getting list of included headers: " +
+                               output.standard_error());
     }
 
   private: